# dev-docs/openapi-mvp.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "OSMU MVP API",
    "version": "0.1.0-prototype",
    "description": "Machine-readable contract snapshot for the current OSMU prototype. It covers the REST control plane and the MVP S3-compatible alias surface used by the local release gate."
  },
  "servers": [
    {
      "url": "http://localhost:8080",
      "description": "Local prototype backend"
    }
  ],
...
  "security": [
    { "bearerAuth": [] }
  ],
  "paths": {
...
    "/api/auth/login": {
      "post": {
        "tags": ["Auth"],
        "operationId": "login",
        "summary": "Create access and refresh tokens",
        "security": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": { "$ref": "#/components/schemas/LoginRequest" }
            }
          }
        },
        "responses": {
          "200": { "$ref": "#/components/responses/JsonOk" },
          "401": { "$ref": "#/components/responses/Error" }
        }
      }
    },
    "/api/auth/refresh": {
      "post": {
        "tags": ["Auth"],
        "operationId": "refreshSession",
        "summary": "Rotate refresh token and issue a new access token",
        "security": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": { "$ref": "#/components/schemas/RefreshTokenRequest" }
            }
          }
        },
        "responses": {
          "200": { "$ref": "#/components/responses/JsonOk" },
          "401": { "$ref": "#/components/responses/Error" }
        }
      }
    },
...
    "/api/access-keys": {
      "get": {
        "tags": ["Access Keys"],
        "operationId": "getAccessKeys",
        "summary": "List access keys",
        "responses": {
          "200": { "$ref": "#/components/responses/JsonList" }
        }
      },
      "post": {
        "tags": ["Access Keys"],
        "operationId": "createAccessKey",
        "summary": "Create an access key and one-time secret",
        "requestBody": { "$ref": "#/components/requestBodies/JsonObject" },
        "responses": {
          "200": { "$ref": "#/components/responses/JsonOk" },
          "201": { "$ref": "#/components/responses/JsonOk" }
        }
      }
    },
    "/api/access-keys/{keyId}": {
      "delete": {
        "tags": ["Access Keys"],
        "operationId": "deleteAccessKey",
        "summary": "Deactivate an access key",
        "parameters": [
          { "$ref": "#/components/parameters/KeyId" }
        ],
        "responses": {
          "204": { "description": "Deactivated" }
        }
      }
    },
...
    "/api/admin/quota-policies/history": {
      "get": {
        "tags": ["Admin"],
        "operationId": "getQuotaPolicyHistory",
        "summary": "List quota policy change history",
        "parameters": [
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": { "type": "integer", "minimum": 1, "maximum": 200, "default": 50 }
          }
        ],
        "responses": {
          "200": { "$ref": "#/components/responses/JsonList" }
        }
      }
    },
    "/api/admin/quota-policies/{targetType}/{targetId}": {
      "put": {
        "tags": ["Admin"],
        "operationId": "saveQuotaPolicy",
        "summary": "Create or update a user, organization, or bucket quota policy",
        "parameters": [
          {
            "name": "targetType",
            "in": "path",
            "required": true,
            "schema": { "type": "string", "enum": ["USER", "ORGANIZATION", "BUCKET"] }
          },
          {
            "name": "targetId",
            "in": "path",
            "required": true,
            "schema": { "type": "integer", "minimum": 1 }
          },
          {
            "name": "reason",
            "in": "query",
            "required": false,
            "schema": { "type": "string", "maxLength": 512 }
          }
        ],
        "requestBody": { "$ref": "#/components/requestBodies/JsonObject" },
        "responses": {
          "200": { "$ref": "#/components/responses/JsonOk" },
          "400": { "$ref": "#/components/responses/Error" }
        }
      },
      "delete": {
        "tags": ["Admin"],
        "operationId": "deleteQuotaPolicy",
        "summary": "Delete a quota policy",
        "parameters": [
          {
            "name": "targetType",
            "in": "path",
            "required": true,
            "schema": { "type": "string", "enum": ["USER", "ORGANIZATION", "BUCKET"] }
          },
          {
            "name": "targetId",
            "in": "path",
            "required": true,
            "schema": { "type": "integer", "minimum": 1 }
          }
        ],
        "responses": {
          "204": { "description": "Deleted" },
          "404": { "$ref": "#/components/responses/Error" }
        }
      }
    },
    "/api/admin/object-share-policy": {
      "get": {
        "tags": ["Admin"],
        "operationId": "getObjectSharePolicy",
        "summary": "Read global object share link policy",
        "responses": {
          "200": { "$ref": "#/components/responses/JsonOk" }
        }
      },
      "put": {
        "tags": ["Admin"],
        "operationId": "saveObjectSharePolicy",
        "summary": "Update global object share link policy",
        "requestBody": { "$ref": "#/components/requestBodies/JsonObject" },
        "responses": {
          "200": { "$ref": "#/components/responses/JsonOk" },
          "400": { "$ref": "#/components/responses/Error" }
        }
      }
    },
    "/api/admin/object-share-analytics": {
      "get": {
        "tags": ["Admin"],
        "operationId": "getObjectShareAnalytics",
        "summary": "Read global object share link analytics",
        "parameters": [
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": { "type": "integer", "minimum": 1, "maximum": 50, "default": 10 }
          },
          {
            "name": "bucketName",
            "in": "query",
            "required": false,
            "schema": { "type": "string" }
          },
          {
            "name": "status",
            "in": "query",
            "required": false,
            "schema": { "type": "string", "enum": ["ACTIVE", "EXPIRED", "REVOKED", "LIMIT_REACHED"] }
          }
        ],
        "responses": {
          "200": { "$ref": "#/components/responses/JsonOk" },
          "400": { "$ref": "#/components/responses/Error" }
        }
      }
    },
...
  },
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      },
      "osmuAccessKey": {
        "type": "apiKey",
        "in": "header",
        "name": "X-OSMU-Access-Key"
      },
      "awsSigV4": {
        "type": "apiKey",
        "in": "header",
        "name": "Authorization",
        "description": "AWS Signature Version 4 authorization header or equivalent presigned query parameters."
      }
    },
    "parameters": {
...
      "KeyId": {
        "name": "keyId",
        "in": "path",
        "required": true,
        "schema": { "type": "string" }
      },
...
    },
    "requestBodies": {
      "JsonObject": {
        "required": true,
        "content": {
          "application/json": {
            "schema": { "type": "object", "additionalProperties": true }
          }
        }
      },
...
    },
    "responses": {
      "JsonOk": {
        "description": "JSON response",
        "content": {
          "application/json": {
            "schema": { "type": "object", "additionalProperties": true }
          }
        }
      },
      "JsonList": {
        "description": "JSON list response",
        "content": {
          "application/json": {
            "schema": { "$ref": "#/components/schemas/ListResponse" }
          }
        }
      },
...
      "Error": {
        "description": "OSMU JSON error response",
        "content": {
          "application/json": {
            "schema": { "$ref": "#/components/schemas/ErrorResponse" }
          }
        }
      },
...
    },
    "schemas": {
      "LoginRequest": {
        "type": "object",
        "required": ["loginId", "password"],
        "properties": {
          "loginId": { "type": "string" },
          "password": { "type": "string", "format": "password" }
        }
      },
      "RefreshTokenRequest": {
        "type": "object",
        "required": ["refreshToken"],
        "properties": {
          "refreshToken": { "type": "string" }
        }
      },
...
      "ListResponse": {
        "type": "object",
        "properties": {
          "items": {
            "type": "array",
            "items": { "type": "object", "additionalProperties": true }
          },
          "nextCursor": { "type": "string", "nullable": true }
        }
      },
      "ErrorResponse": {
        "type": "object",
        "properties": {
          "code": { "type": "string" },
          "message": { "type": "string" },
          "requestId": { "type": "string" }
        }
      }
    }
  }
}